feat: modularize and organize whole code
