+++ src/config/swagger.js
@@ -1,0 +1,35 @@
+const swaggerJsdoc = require("swagger-jsdoc");
+const swaggerUi = require("swagger-ui-express");
+
+const swaggerOptions = {
+  definition: {
+    openapi: "3.0.0",
+    info: {
+      title: "Translation API",
+      version: "1.0.0",
+      description: "API for translating text using Google Translate",
+    },
+    components: {
+      securitySchemes: {
+        bearerAuth: {
+          type: "http",
+          scheme: "bearer",
+          bearerFormat: "JWT",
+        },
+      },
+    },
+    security: [
+      {
+        bearerAuth: [],
+      },
+    ],
+  },
+  apis: ["src/routes/*.js"], // Adjust path for Swagger docs
+};
+
+const setupSwagger = (app) => {
+  const swaggerDocs = swaggerJsdoc(swaggerOptions);
+  app.use("/docs", swaggerUi.serve, swaggerUi.setup(swaggerDocs));
+};
+
+module.exports = setupSwagger;
